+++ CHANGELOG.md
@@ -23,7 +23,7 @@ Changelog
 
 * Added restriction for internet administration (*inet_admin*)
 * Added restriction for internet bandwidth statistics/administration (*inet_bw*)
-* Added restriction for [VPN]((http://en.wikipedia.org/wiki/Vpn) services (*inet_vpn*)
+* Added restriction for [VPN](http://en.wikipedia.org/wiki/Vpn) services (*inet_vpn*)
 * Added restriction for [Mesh networking](http://en.wikipedia.org/wiki/Mesh_networking) services (*inet_mesh*)
 * Added restriction for [MTP](http://en.wikipedia.org/wiki/Media_Transfer_Protocol)
 * Fixed not disabling on demand function settings when on demand restricting for category is disabled